+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Move `asInstanceOf(ClassElement)` to `DartType`, so that it is also
   supported for `TypeParameterType` when its bound implements the
   requested interface.
+* Fixed invalid implicit downcast in `InterfaceType.allSupertypes`.
 
 ## 0.39.14
 * Removed `defaultSdkDirectory()` and `getSdkProperty()` from internal